% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/suma.dos.numeros.R
\name{suma.dos.numeros}
\alias{suma.dos.numeros}
\title{suma.dos.numeros Function}
\usage{
suma.dos.numeros(a, b)
}
\arguments{
\item{value}{a and value b}
}
\description{
This function allows you sum to numbers
}
\examples{
suma.dos.numeros()
}
\keyword{suma.dos.numeros}

